python/requirements.txt
gensim==3.6.0
pandas==0.25.3
numpy==1.16
tqdm==4.28.1
scipy==1.1.0
pymystem3==0.2.0
nltk==3.4
annoy==1.15.1
lightgbm==2.2.1
jupyter==1.0.0
matplotlib==2.2.2
torch==1.2.0
torchtext==0.3.1
torchvision==0.3.0
pytorch-ignite==0.2.0
SQLAlchemy==1.3.11
psycopg2==2.8.4
tensorboardX==1.8
scikit-learn==0.22
pyemd==0.5.1
python-Levenshtein==0.12.0
requests==2.22.0
